requirements.txt
requests
pandas
pyarrow
python-dotenv
psycopg2-binary